refactor : InputView 수정
